Renamed ProcessRequest function to GameEngine.  Also, it didn't make sense for the client to create a channel to handle the responses, so I used the one that is in the GameManager object instead.  This also means that we can remove GameResponse from the GameRequest object.
